Create plot_offset_delay.py
add python tool to plot the statistics file 1_ppt.tx
ptpd2 -s -i eth0 -L -S /data/1_ptp.txt

you need to  delete the header in the file and the other unrelated line in the file.  you can run python and zoom the figure and modify the code to get the percentage of the offset less than 500us.